# core/openapi.yaml
openapi: 3.1.0
info:
  title: Kobo API
  version: 0.1.0
  summary: Identity-anchored virtual account infrastructure built on Monnify.
  description: >
    Kobo provisions and manages persistent, identity-linked virtual accounts.
    Integrators register identities, receive a dedicated virtual account per
    identity, and consume reconciled transactions, statements, and lifecycle
    events through this API. Integrators never call Monnify directly.

    Conventions for implementers:
    - All monetary amounts are integers in kobo (the smallest Naira unit), never floats.
    - All timestamps are ISO 8601 / RFC 3339 in UTC, e.g. "2026-06-30T14:00:00Z".
    - All resource IDs are UUIDv4 strings.
    - Pagination uses cursor-based `page[cursor]` and `page[limit]` query params, response includes `next_cursor` (nullable).
    - Every error response uses the shared Error schema with a stable `code` field; integrators should branch on `code`, not on the human-readable `message`.
  contact:
    name: Kobo Platform Team

servers:
  - url: https://api.kobo.triumphsystems.tech/v1
    description: Production
  - url: https://sandbox.api.kobo.triumphsystems.tech/v1
    description: Sandbox (mirrors production behavior, including duplicate webhook delivery)

security:
  - ApiKeyAuth: []
    HmacSignature: []

tags:
  - name: Identities
    description: Registering and managing the root identity object and its lifecycle.
  - name: Accounts
    description: Reading transaction history and statements for an identity's virtual account.
  - name: Exceptions
    description: Misdirected payments and unmatched transfers requiring resolution.
  - name: Webhooks
    description: Inbound endpoint Monnify calls; documented for completeness, not called by integrators.
  - name: System
    description: Health and meta endpoints.

paths:
  /identities:
    get:
      tags: [Identities]
      operationId: listIdentities
      summary: List all identities for the integrator.
      parameters:
        - name: state
          in: query
          description: Filter identities by state.
          required: false
          schema:
            $ref: '#/components/schemas/IdentityState'
        - name: limit
          in: query
          description: Max number of records to return.
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: offset
          in: query
          description: Number of records to skip.
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: A list of identities.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Identity'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Identities]
      operationId: createIdentity
      summary: Register a new identity and provision its virtual account.
      description: >
        Creates an Identity record and synchronously provisions a Monnify virtual account
        against it. The response typically returns immediately with the identity in the ACTIVE
        state and the virtual account details populated. If the downstream provider is temporarily
        unavailable, the identity creation is rolled back and an error is returned.
        can poll GET /identities/{identityId} or listen for the `identity.activated` webhook.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIdentityRequest'
            examples:
              student:
                value:
                  external_reference: "school_lautech_001:student:4471"
                  metadata:
                    class: "JSS 2"
                    school_id: "school_lautech_001"
      responses:
        '201':
          description: Identity created and provisioning started.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Identity'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: An identity with this external_reference already exists for this integrator.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: duplicate_external_reference
                message: identity with external reference already exists
        '502':
          description: Downstream provider unavailable. Identity creation was rolled back.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: provisioning_failed
                message: failed to create virtual account with provider

  /identities/{identityId}:
    parameters:
      - $ref: '#/components/parameters/IdentityId'
    get:
      tags: [Identities]
      operationId: getIdentity
      summary: Fetch an identity record, its current state, and linked account.
      responses:
        '200':
          description: Identity found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Identity'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Identities]
      operationId: updateIdentity
      summary: Update display profile fields (rename, contact details).
      description: >
        Updates only the display profile layer. Never changes the identity ID
        and never triggers a new virtual account provisioning. If
        display_name changes, Kobo re-submits the updated name to Monnify for
        the linked virtual account and logs an identity event of type
        "renamed".
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateIdentityRequest'
      responses:
        '200':
          description: Identity updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Identity'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Identity is in CLOSED or CLOSING state and cannot be updated. Reopen first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /identities/{identityId}/close:
    parameters:
      - $ref: '#/components/parameters/IdentityId'
    post:
      tags: [Identities]
      operationId: closeIdentity
      summary: Initiate closure of an identity's virtual account.
      description: >
        Transitions the account ACTIVE or LIMITED -> CLOSING. The account
        stops accepting new attributed inflow immediately; any settlement
        already in flight is allowed to clear. Once balance reaches zero and
        funds are swept to sweep_destination, the account transitions
        CLOSING -> CLOSED automatically and an `identity.closed` webhook
        fires. This endpoint does not block until closure completes; it
        starts the process.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CloseIdentityRequest'
      responses:
        '202':
          description: Closure initiated, account now in CLOSING state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Identity'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Identity is not in a state that can transition to CLOSING (e.g. already CLOSED or already CLOSING).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /identities/{identityId}/reopen:
    parameters:
      - $ref: '#/components/parameters/IdentityId'
    post:
      tags: [Identities]
      operationId: reopenIdentity
      summary: Reopen a CLOSED identity's account.
      description: >
        Transitions CLOSED -> ACTIVE against the same identity ID, without
        re-running full provisioning. Fails if the identity is not currently
        CLOSED.
      responses:
        '200':
          description: Identity reopened.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Identity'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Identity is not in CLOSED state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /accounts/{accountId}/transactions:
    parameters:
      - $ref: '#/components/parameters/AccountId'
      - $ref: '#/components/parameters/PageCursor'
      - $ref: '#/components/parameters/PageLimit'
    get:
      tags: [Accounts]
      operationId: listAccountTransactions
      summary: Paginated, reconciled transaction history for an account.
      responses:
        '200':
          description: Transaction list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /accounts/{accountId}/statement:
    parameters:
      - $ref: '#/components/parameters/AccountId'
      - name: period
        in: query
        description: 'Statement period in YYYY-MM format. Defaults to the current month if omitted.'
        required: false
        schema:
          type: string
          pattern: '^\d{4}-\d{2}$'
          example: "2026-06"
    get:
      tags: [Accounts]
      operationId: getAccountStatement
      summary: Structured statement for a given period.
      responses:
        '200':
          description: Statement.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Statement'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /exceptions:
    get:
      tags: [Exceptions]
      operationId: listExceptions
      summary: List unresolved misdirected-payment or unmatched-transfer cases.
      parameters:
        - $ref: '#/components/parameters/PageCursor'
        - $ref: '#/components/parameters/PageLimit'
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [open, resolved]
            default: open
      responses:
        '200':
          description: Exception list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /exceptions/{exceptionId}/resolve:
    parameters:
      - name: exceptionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Exceptions]
      operationId: resolveException
      summary: Apply a resolution to a flagged exception.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveExceptionRequest'
      responses:
        '200':
          description: Exception resolved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Exception is already resolved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /webhooks/monnify:
    post:
      tags: [Webhooks]
      operationId: receiveMonnifyWebhook
      summary: Internal endpoint Monnify calls into. Not used by Kobo integrators.
      description: >
        Documented for completeness and for engineers implementing the
        reconciliation engine. Verifies the Monnify signature header before
        processing. Every event includes a transaction reference used as the
        idempotency key; duplicate deliveries are accepted (200) but produce
        no additional ledger writes.
      security:
        - MonnifyWebhookSignature: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MonnifyWebhookPayload'
      responses:
        '200':
          description: Event accepted (processed or deduplicated).
        '400':
          description: Malformed payload.
        '401':
          description: Invalid or missing Monnify signature.

  /healthz:
    get:
      tags: [System]
      operationId: healthCheck
      summary: Liveness/readiness check.
      security: []
      responses:
        '200':
          description: Service healthy.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    const: ok
                  db:
                    type: string
                    enum: [ok, degraded]

components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: basic
      description: Basic Auth using API Key as the username and API Secret as the password.
    HmacSignature:
      type: apiKey
      in: header
      name: X-Kobo-Signature
      description: >
        HMAC-SHA256 signature of the request body plus the X-Kobo-Timestamp
        header value, signed with the integrator's API secret. Requests with
        a timestamp older than 5 minutes are rejected to prevent replay.
    MonnifyWebhookSignature:
      type: apiKey
      in: header
      name: monnify-signature
      description: Signature provided by Monnify on outbound webhook calls, verified against the shared webhook secret.

  parameters:
    IdentityId:
      name: identityId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    AccountId:
      name: accountId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageCursor:
      name: page[cursor]
      in: query
      required: false
      schema:
        type: string
    PageLimit:
      name: page[limit]
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50

  responses:
    BadRequest:
      description: Request failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid API key / signature.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found, or not visible to this integrator's namespace.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  schemas:
    IdentityState:
      type: string
      description: >
        Lifecycle state of the identity's virtual account. See
        docs/LIFECYCLE.md for the full transition table.
      enum:
        - pending
        - active
        - limited
        - closing
        - closed

    Identity:
      type: object
      required: [id, external_reference, display_name, state, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        external_reference:
          type: string
          description: Integrator-supplied stable reference for this identity (e.g. their own internal customer ID). Unique per integrator.
        display_name:
          type: string
        state:
          $ref: '#/components/schemas/IdentityState'
        virtual_account:
          $ref: '#/components/schemas/VirtualAccountSummary'
        metadata:
          type: object
          additionalProperties: true
          description: Arbitrary integrator-supplied key-value data, returned as-is.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    VirtualAccountSummary:
      type: object
      nullable: true
      description: Null while the identity is in "pending" or "failed" state.
      required: [account_number, bank_name, account_name]
      properties:
        account_number:
          type: string
          example: "1234567890"
        bank_name:
          type: string
          example: "Monnify MFB"
        account_name:
          type: string
          example: "John Ade"
        expected_amount_kobo:
          type: integer
          nullable: true
          description: If set, the virtual account will only accept payments matching this exact amount.
          example: 1500000
        is_expired:
          type: boolean
          description: Whether the virtual account has been explicitly expired.
          default: false
          example: false

    CreateIdentityRequest:
      type: object
      required: [external_reference, display_name]
      properties:
        external_reference:
          type: string
          minLength: 1
          maxLength: 200
        display_name:
          type: string
          minLength: 8
          maxLength: 64
        metadata:
          type: object
          additionalProperties: true

    UpdateIdentityRequest:
      type: object
      description: At least one field must be present.
      properties:
        display_name:
          type: string
          minLength: 8
          maxLength: 64
        metadata:
          type: object
          additionalProperties: true

    CloseIdentityRequest:
      type: object
      required: [sweep_destination]
      properties:
        sweep_destination:
          $ref: '#/components/schemas/SweepDestination'
        reason:
          type: string
          maxLength: 500

    SweepDestination:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [refund_to_source, integrator_account, successor_identity]
        successor_identity_id:
          type: string
          format: uuid
          description: Required when type is "successor_identity".
        integrator_account_reference:
          type: string
          description: Required when type is "integrator_account"; integrator-supplied bank account reference on file.

    Transaction:
      type: object
      required: [id, account_id, amount_kobo, direction, status, monnify_reference, occurred_at]
      properties:
        id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        amount_kobo:
          type: integer
          description: Amount in kobo (1/100 Naira). Always positive; direction indicates flow.
        direction:
          type: string
          enum: [inbound]
          description: v1 only supports inbound attribution; outbound transfers are out of scope for this primitive.
        status:
          type: string
          enum: [matched, partial, overpayment]
        monnify_reference:
          type: string
          description: Monnify's transaction reference, used as the idempotency key.
        occurred_at:
          type: string
          format: date-time

    TransactionListResponse:
      type: object
      required: [data, next_cursor]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
        next_cursor:
          type: string
          nullable: true

    Statement:
      type: object
      required: [account_id, period, opening_balance_kobo, closing_balance_kobo, total_inflow_kobo, transactions]
      properties:
        account_id:
          type: string
          format: uuid
        period:
          type: string
          example: "2026-06"
        opening_balance_kobo:
          type: integer
        closing_balance_kobo:
          type: integer
        total_inflow_kobo:
          type: integer
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'

    ExceptionType:
      type: string
      enum:
        - payment_to_closed_account
        - payment_to_unknown_account
        - payment_during_closing

    Exception:
      type: object
      required: [id, type, amount_kobo, monnify_reference, status, detected_at]
      properties:
        id:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/ExceptionType'
        amount_kobo:
          type: integer
        monnify_reference:
          type: string
        related_account_id:
          type: string
          nullable: true
          format: uuid
        status:
          type: string
          enum: [open, resolved]
        resolution:
          $ref: '#/components/schemas/ExceptionResolution'
        detected_at:
          type: string
          format: date-time
        resolved_at:
          type: string
          nullable: true
          format: date-time

    ExceptionResolution:
      type: object
      nullable: true
      properties:
        action:
          type: string
          enum: [return_to_sender, redirect_to_successor, manual_override]
        notes:
          type: string

    ResolveExceptionRequest:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [return_to_sender, redirect_to_successor, manual_override]
        successor_identity_id:
          type: string
          format: uuid
          description: Required when action is "redirect_to_successor".
        notes:
          type: string
          maxLength: 500

    ExceptionListResponse:
      type: object
      required: [data, next_cursor]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Exception'
        next_cursor:
          type: string
          nullable: true

    MonnifyWebhookPayload:
      type: object
      description: Shape as received from Monnify; field names follow Monnify's contract.
      required: [event_type, requestId, data]
      properties:
        event_type:
          type: string
          example: "payment_success"
        requestId:
          type: string
        data:
          type: object
          properties:
            merchant:
              type: object
              properties:
                userId:
                  type: string
                walletId:
                  type: string
            transaction:
              type: object
              properties:
                transactionId:
                  type: string
                type:
                  type: string
                transactionAmount:
                  type: number
                  description: Amount in Naira as sent by Monnify.
                time:
                  type: string
                  format: date-time
                responseCode:
                  type: string
                aliasAccountNumber:
                  type: string
            customer:
              type: object

    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable error code, e.g. "identity_not_found", "invalid_transition", "duplicate_external_reference".
        message:
          type: string
          description: Human-readable description, not stable, do not branch on this.
        details:
          type: object
          additionalProperties: true